made parameters a list again in Feature resource
A map doesn't make sense from a RDF point of view here, only within the
range field. From a JSON point of view it does make sense though, but as
JSON-LD doesn't allow to map a map to a list in its context we just bite
the bullet and accept this minor inconvenience.